Split report generator journal logic
